Refactor: AnswerPage 폴더 정리
